llm/requirements.txt
--find-links https://download.pytorch.org/whl/cpu
numpy
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
openai>=1.0.0
python-dotenv>=1.0.0
groq>=0.3.0
